fix and enhancement according review suggestions
