+++ client/src/modules/ward/configuration/bed/bed.html
