+++ client/src/modules/ward/configuration/bed/bed.service.js
