+++ client/src/modules/ward/configuration/bed/modals/createUpdate.html
