+++ client/src/modules/ward/configuration/bed/templates/action.tmpl.html
