+++ client/src/modules/ward/configuration/configuration.html
