+++ client/src/modules/ward/configuration/room/room.html
