+++ client/src/modules/ward/configuration/room/templates/action.tmpl.html
