+++ client/src/modules/ward/configuration/ward/templates/action.tmpl.html
