+++ client/src/modules/ward/configuration/ward/ward.service.js
