+++ client/src/modules/ward/ward_module.routes.js
@@ -4,21 +4,21 @@ angular.module('bhima.routes')
 function wardModuleStateProvider($stateProvider) {
   $stateProvider
     .state('wardModuleConfiguration', {
-      url : '/ward_module/configuration',
+      url : '/ward/configuration',
       views : {
         '' : {
-          templateUrl : 'modules/ward_module/configuration/configuration.html',
+          templateUrl : 'modules/ward/configuration/configuration.html',
         },
         'ward@wardModuleConfiguration' : {
-          templateUrl : 'modules/ward_module/configuration/ward/ward.html',
+          templateUrl : 'modules/ward/configuration/ward/ward.html',
           controller : 'WardController as WardCtrl',
         },
         'room@wardModuleConfiguration' : {
-          templateUrl : 'modules/ward_module/configuration/room/room.html',
+          templateUrl : 'modules/ward/configuration/room/room.html',
           controller : 'RoomController as RoomCtrl',
         },
         'bed@wardModuleConfiguration' : {
-          templateUrl : 'modules/ward_module/configuration/bed/bed.html',
+          templateUrl : 'modules/ward/configuration/bed/bed.html',
           controller : 'BedController as BedCtrl',
         },
       },
