+++ test/end-to-end/ward/bed/bed.spec.js
@@ -8,7 +8,7 @@ const page = new BedPage();
 function BedManagementTests() {
 
   // navigate to the page
-  before(() => helpers.navigate('#/ward_module/configuration'));
+  before(() => helpers.navigate('#/ward/configuration'));
 
   const bed = 'PA.RA.001';
