+++ test/end-to-end/ward/ward/ward.spec.js
@@ -9,7 +9,7 @@ const page = new WardPage();
 function WardManagementTests() {
 
   // navigate to the page
-  before(() => helpers.navigate('#/ward_module/configuration'));
+  before(() => helpers.navigate('#/ward/configuration'));
 
   it('should add a new Ward', () => {
     page.openCreateModal();
